+++ Winium/Winium.StoreApps.Driver/CommandExecutors/GetWindowSizeExecutor.cs
@@ -14,7 +14,6 @@
 
         protected override string DoImpl()
         {
-            this.Automator.UpdatedOrientationForEmulatorController();
             var phoneScreenSize = this.Automator.EmulatorController.PhoneScreenSize;
 
             return this.JsonResponse(