+++ Winium/Winium.StoreApps.Driver/CommandExecutors/QuitExecutor.cs
@@ -9,6 +9,7 @@
             if (!this.Automator.ActualCapabilities.DebugConnectToRunningApp)
             {
                 // TODO quit should close all open windows (apps) and possible close the emulator
+                this.Automator.EmulatorController.Disconnect();
                 this.Automator.Deployer.Uninstall();
             }