+++ Winium/Winium.StoreApps.Driver/CommandExecutors/TouchFlickExecutor.cs
@@ -17,8 +17,6 @@
 
         protected override string DoImpl()
         {
-            this.Automator.UpdatedOrientationForEmulatorController();
-
             var screen = this.Automator.EmulatorController.PhoneScreenSize;
             var startPoint = new Point(screen.Width / 2, screen.Height / 2);
             var elementId = Automator.GetValue<string>(this.ExecutedCommand.Parameters, "element");